feat: add MyLottoInputController
